refactor(dbaas): simplify service type JSON generating and parsing
